desafio concluido

+++ CartasSuperTrunfo.c
@@ -7,7 +7,7 @@ int main() {
     double area1, pib1;
     int pontos1;
     double densidade1, pib_per_capita1;
-    float super_poder1;
+    double super_poder1;
 
     // Carta 2
     char estado2, codigo2[5], cidade2[50];
@@ -15,7 +15,7 @@ int main() {
     double area2, pib2;
     int pontos2;
     double densidade2, pib_per_capita2;
-    float super_poder2;
+    double super_poder2;
 
     // Entrada Carta 1
     printf("=== Cadastro da Carta 1 ===\n");
@@ -30,7 +30,7 @@ int main() {
     scanf(" %49[^\n]", cidade1);
 
     printf("Populacao: ");
-    scanf("%d", &populacao1);
+    scanf("%lu", &populacao1);
 
     printf("Area (km²): ");
     scanf("%lf", &area1);
@@ -42,15 +42,15 @@ int main() {
     scanf("%d", &pontos1);
 
     // Cálculos Carta 1
-    densidade1 = (area1 > 0) ? (populacao1 / area1) : 0;
+    densidade1 = (area1 > 0) ? ((double)populacao1 / area1) : 0;
     pib_per_capita1 = (populacao1 > 0) ? (pib1 * 1000000000.0) / populacao1 : 0;
     
     super_poder1 =
-        (float)populacao1 +
-        (float)area1 +
-        (float)pib1 +
-        (float)pontos1 +
-        (float)pib_per_capita1 +
+        populacao1 +
+        area1 +
+        pib1 +
+        pontos1 +
+        pib_per_capita1 +
         (densidade1 > 0 ? (float)(1.0 / densidade1) : 0);
 
     // Entrada Carta 2
@@ -66,7 +66,7 @@ int main() {
     scanf(" %49[^\n]", cidade2);
 
     printf("Populacao: ");
-    scanf("%d", &populacao2);
+    scanf("%lu", &populacao2);
 
     printf("Area (km²): ");
     scanf("%lf", &area2);
@@ -78,15 +78,15 @@ int main() {
     scanf("%d", &pontos2);
 
     // Cálculos Carta 2
-    densidade2 = (area2 > 0) ? (populacao2 / area2) : 0;
+    densidade2 = (area2 > 0) ? ((double)populacao2 / area2) : 0;
     pib_per_capita2 = (populacao2 > 0) ? (pib2 * 1000000000.0) / populacao2 : 0;
     
     super_poder2 =
-        (float)populacao2 +
-        (float)area2 +
-        (float)pib2 +
-        (float)pontos2 +
-        (float)pib_per_capita2 +
+        populacao2 +
+        area2 +
+        pib2 +
+        pontos2 +
+        pib_per_capita2 +
         (densidade2 > 0 ? (float)(1.0 / densidade2) : 0);
 
     // Comparações (1 = Carta 1 vence, 0 = Carta 2 vence)
